# src/main/resources/api/openapi.yaml
openapi: 3.0.4
info:
  title: Sample Ecommerce App
  description: "'This is a ***sample ecommerce app API***.  You can find out more\
    \ about Swagger at [swagger.io](http://swagger.io). Description supports markdown\
    \ markup. For example, you can use the `inline code` using back ticks.'\n"
  termsOfService: https://github.com/chrisgya/webflux-r2dbc/blob/master/LICENSE
  contact:
    email: chrisgya500@gmail.com
  license:
    name: MIT
    url: https://github.com/chrisgya/webflux-r2dbc/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://ecommerce.swagger.io/v3
...
paths:
  /api/v1/carts/{customerId}:
    get:
      tags:
        - cart
      summary: Returns the shopping cart
      description: Returns the shopping cart of given customer
      operationId: getCartByCustomerId
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        "404":
          description: Given customer ID doesn't exist
          content: {}
    delete:
      tags:
        - cart
      summary: Delete the shopping cart
      description: Deletes the shopping cart of given customer
      operationId: deleteCart
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "204":
          description: successful operation
        "404":
          description: Given customer ID doesn't exist
          content: {}
  /api/v1/carts/{customerId}/items:
    get:
      tags:
        - cart
      summary: Returns the list of products in user's shopping cart
      description: Returns the items in shopping cart of given customer
      operationId: getCartItemsByCustomerId
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        "404":
          description: Given customer ID doesn't exist
          content: {}
    put:
      tags:
        - cart
      summary: Replace/add an item in shopping cart
      description: "Adds an item to the shopping cart if it doesn't already exist,\
        \ or replace with given item if it does."
      operationId: addOrReplaceItemsByCustomerId
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: Item object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        "201":
          description: Item added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        "404":
          description: Given customer ID doesn't exist
          content: {}
    post:
      tags:
        - cart
      summary: Adds an item in shopping cart
      description: "Adds an item to the shopping cart if it doesn't already exist,\
        \ or increment quantity by the specified number of items if it does."
      operationId: addCartItemsByCustomerId
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: Item object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        "201":
          description: Item added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        "404":
          description: Given customer ID doesn't exist
          content: {}
  /api/v1/carts/{customerId}/items/{itemId}:
    get:
      tags:
        - cart
      summary: Returns given item from user's shopping cart
      description: Returns given item from shopping cart of given customer
      operationId: getCartItemsByItemId
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: itemId
          in: path
          description: Item (product) Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: If item exists in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        "404":
          description: Given item (product) ID doesn't exist
          content: {}
    delete:
      tags:
        - cart
      summary: Delete the item from shopping cart
      description: Deletes the item from shopping cart of given customer
      operationId: deleteItemFromCart
      parameters:
        - name: customerId
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: itemId
          in: path
          description: Item (product) Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "202":
          description: "Accepts the request, regardless of whether the specified item\
            \ exists in the cart or not."
...
  /api/v1/customers/{id}:
    get:
      tags:
        - customer
      summary: Returns a customer
      description: Returns a customers identifiable by given ID
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - customer
      summary: Deletes the customer
      description: Deletes the customer identifiable by given ID
      operationId: deleteCustomerById
      parameters:
        - name: id
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "202":
          description: "Request accepted, returns this status even if user does not\
            \ exist"
          content: {}
  /api/v1/customers/{id}/cards:
    get:
      tags:
        - customer
      summary: Returns all customer's cards
      description: Returns all customer's cards based on given customer ID
      operationId: getCardByCustomerId
      parameters:
        - name: id
          in: path
          description: Customer Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
...
  /api/v1/addresses:
    get:
      tags:
        - address
      summary: Returns all user's addresses
      description: "Returns all user's addresses, else empty collection"
      operationId: getAllAddresses
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags:
        - address
      summary: Creates a new user addresses
      description: Creates a new user addresses. Does nothing if address already exists.
      operationId: createAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAddressReq'
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /api/v1/addresses/{id}:
    get:
      tags:
        - address
      summary: Returns user's address
      description: Returns user's address based on given address ID.
      operationId: getAddressesById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags:
        - address
      summary: Deletes user's address
      description: Deletes user's address based on given address ID.
      operationId: deleteAddressesById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "202":
          description: "Accepts the deletion request and perform deletion. If ID does\
            \ not exist, does nothing."
          content: {}
  /api/v1/cards:
    get:
      tags:
        - card
      summary: Returns all user's cards
      description: "Returns all user's cards, else empty collection"
      operationId: getAllCards
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
    post:
      tags:
        - card
      summary: Creates a new user addresses
      description: Creates a new user addresses. Does nothing if address already exists.
      operationId: registerCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCardReq'
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
  /api/v1/cards/{id}:
    get:
      tags:
        - card
      summary: Returns user's card
      description: Returns user's card based on given card ID.
      operationId: getCardById
      parameters:
        - name: id
          in: path
          description: card Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
    delete:
      tags:
        - card
      summary: Deletes card's address
      description: Deletes card's address based on given card ID.
      operationId: deleteCardById
      parameters:
        - name: id
          in: path
          description: card Identifier
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "202":
          description: "Accepts the deletion request and perform deletion. If ID does\
            \ not exist, does nothing."
          content: {}
...
components:
  schemas:
    Cart:
      type: object
      properties:
        id:
          type: string
          description: Cart Identifier
        customerId:
          type: string
          description: Id of the customer who possesses the cart
        items:
          type: array
          description: Collection of items in cart.
          items:
            $ref: '#/components/schemas/Item'
      description: Shopping Cart of the user
    Item:
      type: object
      properties:
        id:
          type: string
          description: Item Identifier
        quantity:
          type: integer
          description: The item quantity
          format: int32
        unitPrice:
          type: BigDecimal
          description: The item's price per unit
      description: Items in shopping cart
...
    AddAddressReq:
      xml:
        name: AddAddressReq
      allOf:
        - $ref: '#/components/schemas/Address'
        - type: object
          properties:
            userId:
              type: string
    Address:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        number:
          type: string
          description: house of flat number
        residency:
          type: string
          description: Socieity or building name
        street:
          type: string
          description: street name
        city:
          type: string
          description: city name
        state:
          type: string
          description: state name
        country:
          type: string
          description: country name
        pincode:
          type: string
          description: postal code
      xml:
        name: Address
    Card:
      type: object
      properties:
        id:
          type: string
          description: Order identifier
        cardNumber:
          type: string
          description: Card Number
        expires:
          type: string
          description: Expiration date
        cvv:
          type: string
          description: CVV code
        userId:
          type: string
      xml:
        name: Card
    AddCardReq:
      description: Request object for new card registration.
      xml:
        name: AddCardReq
      allOf:
        - $ref: '#/components/schemas/Card'
        - type: object
          properties:
            userId:
              type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: string
      xml:
        name: User
...